ADD: prop tests for add/sub/mul and refactor prop tests
